anvil-python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
aiohttp==3.9.3
    # via python-libmaas
aiosignal==1.3.1
    # via aiohttp
anyio==4.3.0
    # via httpx
argcomplete==3.2.3
    # via python-libmaas
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
bcrypt==4.1.2
    # via paramiko
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cffi==1.16.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via sunbeam
colorclass==2.2.2
    # via python-libmaas
croniter==2.0.3
    # via sunbeam
cryptography==42.0.5
    # via paramiko
dnspython==2.6.1
    # via pymongo
exceptiongroup==1.2.0
    # via anyio
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.42
    # via sunbeam
google-auth==2.29.0
    # via kubernetes
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via lightkube
hvac==2.1.0
    # via juju
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonschema==4.21.1
    # via sunbeam
jsonschema-specifications==2023.12.1
    # via jsonschema
juju==3.3.1.1
    # via sunbeam
kubernetes==29.0.0
    # via juju
lightkube==0.15.2
    # via sunbeam
lightkube-models==1.29.0.7
    # via
    #   lightkube
    #   sunbeam
macaroonbakery==1.3.4
    # via
    #   juju
    #   python-libmaas
    #   sunbeam
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
netifaces==0.11.0
    # via sunbeam
oauthlib==3.2.2
    # via
    #   kubernetes
    #   python-libmaas
    #   requests-oauthlib
packaging==24.0
    # via juju
paramiko==3.4.0
    # via juju
pbr==6.0.0
    # via sunbeam
petname==2.6
    # via sunbeam
pexpect==4.9.0
    # via sunbeam
protobuf==5.26.0
    # via macaroonbakery
ptyprocess==0.7.0
    # via pexpect
pwgen==0.8.2.post0
    # via sunbeam
pyasn1==0.5.1
    # via
    #   juju
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==1.10.14
    # via
    #   anvil (pyproject.toml)
    #   sunbeam
pygments==2.17.2
    # via rich
pymacaroons==0.13.0
    # via macaroonbakery
pymongo==4.6.2
    # via python-libmaas
pynacl==1.5.0
    # via
    #   macaroonbakery
    #   paramiko
    #   pymacaroons
pyrfc3339==1.1
    # via
    #   juju
    #   macaroonbakery
pyroute2==0.7.12
    # via sunbeam
python-dateutil==2.9.0.post0
    # via
    #   croniter
    #   kubernetes
python-libmaas==0.6.8
    # via sunbeam
pytz==2024.1
    # via
    #   croniter
    #   pyrfc3339
    #   python-libmaas
pyyaml==6.0.1
    # via
    #   juju
    #   kubernetes
    #   lightkube
    #   python-libmaas
    #   snap-helpers
    #   sunbeam
referencing==0.34.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   hvac
    #   kubernetes
    #   macaroonbakery
    #   requests-oauthlib
    #   requests-unixsocket
    #   sunbeam
requests-oauthlib==1.4.0
    # via kubernetes
requests-unixsocket==0.3.0
    # via sunbeam
rich==13.7.1
    # via
    #   anvil (pyproject.toml)
    #   sunbeam
rpds-py==0.18.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   kubernetes
    #   macaroonbakery
    #   pymacaroons
    #   python-dateutil
smmap==5.0.1
    # via gitdb
snap-helpers @ git+https://github.com/skatsaounis/snap-helpers
    # via
    #   anvil (pyproject.toml)
    #   sunbeam
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sunbeam @ git+https://github.com/canonical/snap-openstack@anvil-0.1#subdirectory=sunbeam-python
    # via anvil (pyproject.toml)
terminaltables==3.1.10
    # via python-libmaas
toposort==1.10
    # via juju
typing-extensions==4.10.0
    # via
    #   anyio
    #   pydantic
    #   typing-inspect
typing-inspect==0.9.0
    # via juju
urllib3==1.26.18
    # via
    #   kubernetes
    #   requests
    #   sunbeam
websocket-client==1.7.0
    # via kubernetes
websockets==12.0
    # via juju
yarl==1.9.4
    # via aiohttp
